Simplify package structure and delete unnecessary interfaces.
