+++ trakt-api/implementation/build.gradle.kts
@@ -43,5 +43,6 @@ kotlin {
 }
 
 addKspDependencyForAllTargets(libs.kotlinInject.compiler)
+addKspDependencyForAllTargets(libs.kotlinInject.anvil.compiler)
 
 android { namespace = "com.thomaskioko.trakt.api.implementation" }
