+++ trakt-api/implementation/src/androidMain/kotlin/com/thomaskioko/trakt/service/implementation/TraktPlatformComponent.kt
@@ -1,6 +1,5 @@
-package com.thomaskioko.trakt.service.implementation.inject
+package com.thomaskioko.trakt.service.implementation
 
-import com.thomaskioko.trakt.service.implementation.TraktAuthInterceptor
 import io.ktor.client.engine.okhttp.OkHttp
 import me.tatarka.inject.annotations.Provides
 import software.amazon.lastmile.kotlin.inject.anvil.AppScope
@@ -8,10 +7,10 @@ import software.amazon.lastmile.kotlin.inject.anvil.ContributesTo
 import software.amazon.lastmile.kotlin.inject.anvil.SingleIn
 
 @ContributesTo(AppScope::class)
-actual interface TraktPlatformComponent {
+interface TraktPlatformComponent {
 
-  @SingleIn(AppScope::class)
   @Provides
+  @SingleIn(AppScope::class)
   fun provideTraktHttpClientEngine(interceptor: TraktAuthInterceptor): TraktHttpClientEngine =
     OkHttp.create { addInterceptor(interceptor) }
 }
