--- trakt-api/implementation/src/commonMain/kotlin/com/thomaskioko/trakt/service/implementation/inject/TraktComponent.kt
@@ -1,67 +1,0 @@
-package com.thomaskioko.trakt.service.implementation.inject
-
-import com.thomaskioko.trakt.service.implementation.DefaultTraktListRemoteDataSource
-import com.thomaskioko.trakt.service.implementation.DefaultTraktTokenRemoteDataSource
-import com.thomaskioko.trakt.service.implementation.DefaultTraktUserRemoteDataSource
-import com.thomaskioko.trakt.service.implementation.traktHttpClient
-import com.thomaskioko.tvmaniac.core.base.model.Configs
-import com.thomaskioko.tvmaniac.core.logger.KermitLogger
-import com.thomaskioko.tvmaniac.trakt.api.TraktListRemoteDataSource
-import com.thomaskioko.tvmaniac.trakt.api.TraktTokenRemoteDataSource
-import com.thomaskioko.tvmaniac.trakt.api.TraktUserRemoteDataSource
-import io.ktor.client.HttpClient
-import kotlinx.serialization.json.Json
-import me.tatarka.inject.annotations.Provides
-import software.amazon.lastmile.kotlin.inject.anvil.AppScope
-import software.amazon.lastmile.kotlin.inject.anvil.ContributesTo
-import software.amazon.lastmile.kotlin.inject.anvil.SingleIn
-
-typealias TraktHttpClient = HttpClient
-
-typealias TraktJson = Json
-
-@ContributesTo(AppScope::class)
-interface TraktComponent : TraktPlatformComponent {
-
-  @SingleIn(AppScope::class)
-  @Provides
-  fun provideJson(): TraktJson = Json {
-    ignoreUnknownKeys = true
-    prettyPrint = true
-    encodeDefaults = true
-  }
-
-  @SingleIn(AppScope::class)
-  @Provides
-  fun provideHttpClient(
-    configs: Configs,
-    json: TraktJson,
-    httpClientEngine: TraktHttpClientEngine,
-    logger: KermitLogger,
-  ): TraktHttpClient =
-    traktHttpClient(
-      isDebug = configs.isDebug,
-      traktClientId = configs.traktClientId,
-      json = json,
-      httpClientEngine = httpClientEngine,
-      kermitLogger = logger,
-    )
-
-  @SingleIn(AppScope::class)
-  @Provides
-  fun provideTraktListRemoteDataSource(
-    bind: DefaultTraktListRemoteDataSource,
-  ): TraktListRemoteDataSource = bind
-
-  @SingleIn(AppScope::class)
-  @Provides
-  fun provideTraktTokenRemoteDataSource(
-    bind: DefaultTraktTokenRemoteDataSource,
-  ): TraktTokenRemoteDataSource = bind
-
-  @SingleIn(AppScope::class)
-  @Provides
-  fun provideTraktUserRemoteDataSource(
-    bind: DefaultTraktUserRemoteDataSource,
-  ): TraktUserRemoteDataSource = bind
-}
