--- trakt-api/implementation/src/commonMain/kotlin/com/thomaskioko/trakt/service/implementation/inject/TraktPlatformComponent.kt
@@ -1,7 +1,0 @@
-package com.thomaskioko.trakt.service.implementation.inject
-
-import io.ktor.client.engine.HttpClientEngine
-
-typealias TraktHttpClientEngine = HttpClientEngine
-
-expect interface TraktPlatformComponent
