+++ trakt-api/implementation/src/iosMain/kotlin/com/thomaskioko/trakt/service/implementation/TraktPlatformComponent.kt
@@ -1,4 +1,4 @@
-package com.thomaskioko.trakt.service.implementation.inject
+package com.thomaskioko.trakt.service.implementation
 
 import io.ktor.client.engine.darwin.Darwin
 import me.tatarka.inject.annotations.Provides
@@ -7,9 +7,9 @@ import software.amazon.lastmile.kotlin.inject.anvil.ContributesTo
 import software.amazon.lastmile.kotlin.inject.anvil.SingleIn
 
 @ContributesTo(AppScope::class)
-actual interface TraktPlatformComponent {
+interface TraktPlatformComponent {
 
-  @SingleIn(AppScope::class)
   @Provides
+  @SingleIn(AppScope::class)
   fun provideTraktHttpClientEngine(): TraktHttpClientEngine = Darwin.create()
 }
