--- trakt-api/implementation/src/jvmMain/kotlin/com/thomaskioko/trakt/service/implementation/inject/TraktPlatformComponent.kt
@@ -1,3 +1,0 @@
-package com.thomaskioko.trakt.service.implementation.inject
-
-actual interface TraktPlatformComponent
