+++ trakt-auth/implementation/src/androidMain/kotlin/com/thomaskioko/tvmaniac/traktauth/implementation/TraktAuthAndroidComponent.kt
@@ -6,7 +6,6 @@ import androidx.core.net.toUri
 import com.thomaskioko.tvmaniac.core.base.annotations.ActivityScope
 import com.thomaskioko.tvmaniac.core.base.model.Configs
 import com.thomaskioko.tvmaniac.traktauth.api.TraktAuthManager
-import com.thomaskioko.tvmaniac.traktauth.api.TraktAuthRepository
 import me.tatarka.inject.annotations.Provides
 import net.openid.appauth.AuthorizationRequest
 import net.openid.appauth.AuthorizationService
@@ -20,8 +19,8 @@ import software.amazon.lastmile.kotlin.inject.anvil.SingleIn
 @ContributesTo(ActivityScope::class)
 interface TraktAuthAndroidComponent {
 
-  @SingleIn(ActivityScope::class)
   @Provides
+  @SingleIn(ActivityScope::class)
   fun provideAuthConfig(): AuthorizationServiceConfiguration {
     return AuthorizationServiceConfiguration(
       Uri.parse("https://trakt.tv/oauth/authorize"),
